Modify permissions on staged subdirectories to allow deletion.
Specifically, the subdirectories need to give write and execute
permissions to the service account running the gobbler backend, so that
it can properly purge old files from the staging directory. It's too
much hassle to try to give permissions specifically to the service
account so we just give write permissions to all users.